finalizando backend

+++ Técnico em Desenvolvimentos de Sistemas/Desenvolvimento de Sistemas/! SPRING/app-cad-pro/src/main/java/com/piske/app_cad_pro/repository/ProdutoRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.piske.app_cad_pro.model.Produto;
 
-public interface ProdutoRepository extends JpaRepository<Produto, Integer>{
+public interface ProdutoRepository extends JpaRepository<Produto, Long>{
     
 }
